[v1.0.0] [ 불필요 파일 삭제 ]
